+++ Changelog.md
@@ -1,7 +1,15 @@
 # Changelog
 
+## Version 1.1
+ *   Using new hook provided by [ozh](http://github.com/ozh).
+ *   Licencing under terms of GPLv2 or later to be compatible with
+     Authorization Manager plugin's licensing terms.
+ *   `rbac.php` taken from nicwaller's Authorization Manager plugin for YOURLS
+     and modified to work with Shibboleth entitlements. Renamed functions
+     and classes to avoid name clash.
+
 ## Version 1.0
- *   code refactored to work as a plugin
+ *   code refactored to work as a plugin.
 
 ## Version 0.5 (unreleased)
- *   hacked yourls to support Shibboleth authentication
+ *   hacked yourls to support Shibboleth authentication.